requirements.txt
Flask==2.3.0
Flask-CORS==4.0.0
google-auth-oauthlib==1.0.0
google-auth-httplib2==0.2.0
google-api-python-client==2.90.0
APScheduler==3.10.0
python-dotenv==1.0.0
